feat: enable decimal quantities across all modules. Replaced parseInt with parseFloat in UI and API. Added step=any to quantity inputs. Updated formatting to show 2 decimal places.

+++ app/api/import/route.ts
@@ -34,7 +34,7 @@ export async function POST(request: Request) {
       });
 
       // Insert Inventory Lot
-      const qty = parseInt(item.الكمية || item.quantity || item.الرصيد) || 0;
+      const qty = parseFloat(item.الكمية || item.quantity || item.الرصيد) || 0;
       if (qty > 0) {
          let expDate = new Date();
          expDate.setFullYear(expDate.getFullYear() + 1); // Default to +1 year if not provided
